requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
flet==0.28.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
oauthlib==3.2.2
psycopg2-binary==2.9.10
repath==0.9.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
